Update build definitions
How PrisonBuild works:

1) Prison build clones it's repository link hard-coded into the JAR
2) Looks for this file and executes the jobs listed.

A job with '-' causes PrisonBuild to clone into the specified repository and executes 'gradle build' in the clone directory
A job with '@' adds a file to both zip outputs (for both platforms)
A job with '!' adds a file to a specific zip. The zip entry name inherits the filename.

You could theoretically customise PrisonBuild for your own uses. If you do, make sure the job order is the same as in this file ('-' first, '!' second and '@' last). You need at least one '!' entry to create the zip file.