Still locked, but new mines GUI could be done
Changelogs:
- Added a new GUI, this could need a whole documentation, I'll tell here in a nutshell how does it works.
First of all, delete your old GuiConfig.yml, you'll need the new one
For now it still under the command /prisonmanager mines, same for ranks /prisonmanager ranks, we'll think about a public command.
You'll need to set up, it depends from you warp plugin, the permission without the warpname inside the GuiConfig.yml, same for the command, BY DEFAULT's configured for essentialsX.
The mine must be of the SAME name of the WARP, for example the A mine need an A warp, not something like warp MineA (this won't work).
This will make the GUI works and also the onClick feature to teleport to the mine.