Minor changes.
Changelogs:
- Minor changes
- Attempt fixing double items drops from inventory bug... failed

NOTE:
I think this's the worse bug I've ever attempted to fix and I'm still unable to fix it.
Reproducing this bug:
- Put something inside a backpack and close it.
- Open again the backpack and take that item, so drag and drop it outside the inventory
It will be dropped with success, but you'll see it popup again in the inventory even if the inventory
is never closing and reopening again, IF YOU TRY TO TAKE IT AND DROP IT OUTSIDE AGAIN IT WON'T
REAPPEAR.
Attempted fixes:
- InventoryClickEvent getSlot - 1, doesn't work (check if click with item on cursor is outside of inv).
- InventoryClickEvent getRawSlot -999 doesn't work (same as before).
- InventoryClickEvent getAction == all InventoryAction for drops doesn't work.
- DropEvent doesn't work, it isn't even triggered.
- InventoyDragEvent, sometimes it gets triggered and others doesn't randomly, and doesn't works anyway.
- InventoryClickEvent getClickedInventory -> Should be null, is like the clickedInventory isn't null
or the InventoryClickEvent isn't triggered.
- Many others I do not even remember now.