Trying everything to fix without success, commenting debug messages and committing my things, backpacks are broken.
Uploading a video showing the bug soon, I will give access to it only to who request it to see the bug and debug messages, anyway there still isn't a fix, this's a damn big bug.
IF you want to see the debug message by yourself, un-comment them please, they're in the BackPacksListeners and SpigotPlayerBackPacksGUI classes.